+++ app/src/main/java/org/simple/clinic/AppDatabase.kt
@@ -18,6 +18,8 @@ import org.simple.clinic.medicalhistory.Answer
 import org.simple.clinic.medicalhistory.MedicalHistory
 import org.simple.clinic.overdue.Appointment
 import org.simple.clinic.overdue.AppointmentCancelReason
+import org.simple.clinic.overdue.callresult.CallResult
+import org.simple.clinic.overdue.callresult.Outcome
 import org.simple.clinic.patient.DeletedReason
 import org.simple.clinic.patient.Gender
 import org.simple.clinic.patient.Patient
@@ -80,13 +82,14 @@ import org.simple.clinic.drugs.search.Answer as DrugAnswer
       MedicalOfficer::class,
       TeleconsultationFacilityMedicalOfficersCrossRef::class,
       TeleconsultRecord::class,
-      Drug::class
+      Drug::class,
+      CallResult::class
     ],
     views = [
       OverdueAppointment::class,
       PatientSearchResult::class
     ],
-    version = 94,
+    version = 95,
     exportSchema = true
 )
 @TypeConverters(
@@ -116,7 +119,8 @@ import org.simple.clinic.drugs.search.Answer as DrugAnswer
     TeleconsultStatus.RoomTypeConverter::class,
     DrugCategory.RoomTypeConverter::class,
     DrugAnswer.RoomTypeConverter::class,
-    DrugFrequency.RoomTypeConverter::class
+    DrugFrequency.RoomTypeConverter::class,
+    Outcome.RoomTypeConverter::class
 )
 abstract class AppDatabase : RoomDatabase() {