+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Bump logback to v1.2.5
 - Bump lint to v30.0.1
 - Change large component corner shape to 8dp
+- Add the `CallResult` database table
 
 ### Changes
 - [In Progress: 19 Aug 2021] UI improvements for medical history screen